# backend/api/workflow/openapi.yaml
openapi: 3.0.0
info:
  title: Workflow Management API
  version: 1.0.0
  description: |
    A comprehensive RESTful API for workflow management, execution, and monitoring.
    
    This API provides complete workflow lifecycle management including:
    - Workflow CRUD operations (create, read, update, delete)
    - Execution management (start, stop, pause, resume)
    - Real-time monitoring (status, logs, metrics)
    - Authentication and authorization
    - Rate limiting and request validation
    
    Built with Express.js, TypeScript, and follows OpenAPI 3.0 specifications.
  contact:
    name: API Support
    email: support@jaquedu.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.jaquedu.com/v1
    description: Production server
  - url: https://staging-api.jaquedu.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /workflows:
    post:
      summary: Create a new workflow
      description: Create a new workflow with steps and configuration
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
            example:
              name: "User Onboarding Workflow"
              description: "Automated user onboarding process"
              steps:
                - name: "Send Welcome Email"
                  type: "email"
                  config:
                    template: "welcome"
                    recipient: "{{user.email}}"
                - name: "Create User Account"
                  type: "script"
                  config:
                    script: "createUser.js"
              variables:
                defaultRole: "user"
              metadata:
                tags: ["onboarding", "automation"]
                priority: "normal"
      responses:
        201:
          description: Workflow created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workflow'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        429:
          $ref: '#/components/responses/RateLimitError'

    get:
      summary: List workflows
      description: Retrieve a paginated list of workflows with filtering options
      tags:
        - Workflows
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - in: query
          name: status
          schema:
            type: array
            items:
              $ref: '#/components/schemas/WorkflowStatus'
          description: Filter by workflow status
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
        - in: query
          name: search
          schema:
            type: string
          description: Search in workflow names and descriptions
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [name, createdAt, updatedAt, status]
            default: createdAt
          description: Sort field
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        200:
          description: List of workflows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkflowListResponse'
  /workflows/{id}:
    get:
      summary: Get workflow by ID
      description: Retrieve a specific workflow by its ID
      tags:
        - Workflows
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Workflow ID
      responses:
        200:
          description: Workflow details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workflow'
        404:
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update workflow
      description: Update an existing workflow
      tags:
        - Workflows
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Workflow ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
      responses:
        200:
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workflow'
        400:
          $ref: '#/components/responses/ValidationError'
        404:
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete workflow
      description: Delete a workflow (only if no active executions)
      tags:
        - Workflows
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Workflow ID
      responses:
        204:
          description: Workflow deleted successfully
        404:
          $ref: '#/components/responses/NotFoundError'
        409:
          description: Workflow has active executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /workflows/{id}/execute:
    post:
      summary: Execute workflow
      description: Start execution of a workflow
      tags:
        - Executions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Workflow ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
            example:
              variables:
                user:
                  email: "user@example.com"
                  name: "John Doe"
              priority: "high"
              timeout: 3600000
      responses:
        202:
          description: Workflow execution started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkflowExecution'
        400:
          description: Workflow cannot be executed
        404:
          $ref: '#/components/responses/NotFoundError'
  /executions:
    get:
      summary: List executions
      description: Retrieve a paginated list of workflow executions
      tags:
        - Executions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - in: query
          name: status
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ExecutionStatus'
          description: Filter by execution status
        - in: query
          name: workflowId
          schema:
            type: string
            format: uuid
          description: Filter by workflow ID
      responses:
        200:
          description: List of executions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExecutionListResponse'
  /executions/{executionId}:
    get:
      summary: Get execution details
      description: Retrieve details of a specific execution
      tags:
        - Executions
      parameters:
        - in: path
          name: executionId
          required: true
          schema:
            type: string
            format: uuid
          description: Execution ID
      responses:
        200:
          description: Execution details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkflowExecution'
        404:
          $ref: '#/components/responses/NotFoundError'
  /executions/{executionId}/pause:
    post:
      summary: Pause execution
      description: Pause a running workflow execution
      tags:
        - Executions
      parameters:
        - in: path
          name: executionId
          required: true
          schema:
            type: string
            format: uuid
          description: Execution ID
      responses:
        200:
          description: Execution paused successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkflowExecution'
        400:
          description: Execution cannot be paused
        404:
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - timestamp
        - requestId
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          description: Response data (present on success)
        error:
          $ref: '#/components/schemas/ApiError'
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        requestId:
          type: string
          description: Unique request identifier
...
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/ApiError'
                      - type: object
                        properties:
                          code:
                            example: "MISSING_AUTH"
                          message:
                            example: "Authentication required - provide Bearer token or API key"
...